feature(authentication): jwt authentication user should receive token on successful signin
