+++ .env
@@ -1,1 +1,2 @@
-MONGOHQ_URL=mongodb://127.0.0.1:27017
+MONGOHQ_URL=mongodb://127.0.0.1:27017/rainier-cfh
+JWT_SECRET=Th0asf5sHDGSMNsdf46
